Replace C native integral types with width specific macros from stdint.h

+++ libmonit/src/io/OutputStream.c
@@ -64,7 +64,7 @@ struct T {
         uchar_t *length;
         bool isclosed;
         int sessionWritten;
-        long long int bytesWritten;
+        int64_t bytesWritten;
         uchar_t buffer[BUFFER_SIZE + 1];
 };
 
@@ -392,7 +392,7 @@ int OutputStream_isClosed(T S) {
 }
 
 
-long long int OutputStream_getBytesWritten(T S) {
+int64_t OutputStream_getBytesWritten(T S) {
         assert(S);
         return S->bytesWritten;
 }
