+++ src/net/Link.c
@@ -96,12 +96,12 @@ static struct {
 
 typedef struct LinkData_T {
 #ifndef __LP64__
-        unsigned long long raw;
+        uint64_t raw;
 #endif
-        unsigned long long last;
-        unsigned long long now;
-        unsigned long long minute[60];
-        unsigned long long hour[24];
+        uint64_t last;
+        uint64_t now;
+        uint64_t minute[60];
+        uint64_t hour[24];
 } LinkData_T;
 
 
@@ -109,12 +109,12 @@ struct T {
         char *object;
         const char *(*resolve)(const char *object); // Resolve Object -> Interface, set during Link_T instantiation by constructor (currently we implement only IPAddress -> Interface lookup)
         struct {
-                unsigned long long last;
-                unsigned long long now;
+                uint64_t last;
+                uint64_t now;
         } timestamp;
         int state;       // State (-1 = N/A, 0 = down, 1 = up)
         int duplex;      // Duplex (-1 = N/A, 0 = half, 1 = full)
-        long long speed; // Speed [bps]
+        int64_t speed; // Speed [bps]
         LinkData_T ipackets;  // Packets received on interface
         LinkData_T ierrors;   // Input errors on interface
         LinkData_T ibytes;    // Total number of octets received
@@ -138,8 +138,8 @@ static void __attribute__ ((destructor)) _destructor() {
 /* --------------------------------------------------------------- Private */
 
 
-static void _updateValue(LinkData_T *data, unsigned long long raw) {
-       unsigned long long value = raw;
+static void _updateValue(LinkData_T *data, uint64_t raw) {
+       uint64_t value = raw;
 #ifndef __LP64__
         if (raw < data->raw)
                 value = data->now + ULONG_MAX + 1ULL - data->raw + raw; // Counter wrapped
@@ -171,7 +171,7 @@ static void _updateValue(LinkData_T *data, unsigned long long raw) {
 #endif
 
 
-static void _resetData(LinkData_T *data, unsigned long long value) {
+static void _resetData(LinkData_T *data, uint64_t value) {
 #ifndef __LP64__
         data->raw = value;
 #endif
@@ -196,15 +196,15 @@ static void _reset(T L) {
 }
 
 
-static unsigned long long _deltaSecond(T L, LinkData_T *data) {
+static uint64_t _deltaSecond(T L, LinkData_T *data) {
         if (L->timestamp.last > 0 && L->timestamp.now > L->timestamp.last)
                 if (data->last > 0 && data->now > data->last)
-                        return (unsigned long long)((data->now - data->last) * 1000. / (L->timestamp.now - L->timestamp.last));
+                        return (uint64_t)((data->now - data->last) * 1000. / (L->timestamp.now - L->timestamp.last));
         return 0ULL;
 }
 
 
-static unsigned long long _deltaMinute(T L, LinkData_T *data, int count) {
+static uint64_t _deltaMinute(T L, LinkData_T *data, int count) {
         int stop = Time_minutes(L->timestamp.now / 1000.);
         int delta = stop - count;
         int start = delta < 0 ? 60 + delta : delta;
@@ -216,7 +216,7 @@ static unsigned long long _deltaMinute(T L, LinkData_T *data, int count) {
 }
 
 
-static unsigned long long _deltaHour(T L, LinkData_T *data, int count) {
+static uint64_t _deltaHour(T L, LinkData_T *data, int count) {
         int stop = Time_hour(L->timestamp.now / 1000.);
         int delta = stop - count;
         int start = delta < 0 ? 24 + delta : delta;
@@ -359,25 +359,25 @@ void Link_update(T L) {
 }
 
 
-long long Link_getBytesInPerSecond(T L) {
+int64_t Link_getBytesInPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->ibytes)) : -1LL;
 }
 
 
-long long Link_getBytesInPerMinute(T L, int count) {
+int64_t Link_getBytesInPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->ibytes), count) : -1LL;
 }
 
 
-long long Link_getBytesInPerHour(T L, int count) {
+int64_t Link_getBytesInPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->ibytes), count) : -1LL;
 }
 
 
-long long Link_getBytesInTotal(T L) {
+int64_t Link_getBytesInTotal(T L) {
         assert(L);
         return L->state > 0 ? L->ibytes.now : -1LL;
 }
@@ -389,73 +389,73 @@ double Link_getSaturationInPerSecond(T L) {
 }
 
 
-long long Link_getPacketsInPerSecond(T L) {
+int64_t Link_getPacketsInPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->ipackets)) : -1LL;
 }
 
 
-long long Link_getPacketsInPerMinute(T L, int count) {
+int64_t Link_getPacketsInPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->ipackets), count) : -1LL;
 }
 
 
-long long Link_getPacketsInPerHour(T L, int count) {
+int64_t Link_getPacketsInPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->ipackets), count) : -1LL;
 }
 
 
-long long Link_getPacketsInTotal(T L) {
+int64_t Link_getPacketsInTotal(T L) {
         assert(L);
         return L->state > 0 ? L->ipackets.now : -1LL;
 }
 
 
-long long Link_getErrorsInPerSecond(T L) {
+int64_t Link_getErrorsInPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->ierrors)) : -1LL;
 }
 
 
-long long Link_getErrorsInPerMinute(T L, int count) {
+int64_t Link_getErrorsInPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->ierrors), count) : -1LL;
 }
 
 
-long long Link_getErrorsInPerHour(T L, int count) {
+int64_t Link_getErrorsInPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->ierrors), count) : -1LL;
 }
 
 
-long long Link_getErrorsInTotal(T L) {
+int64_t Link_getErrorsInTotal(T L) {
         assert(L);
         return L->state > 0 ? L->ierrors.now : -1LL;
 }
 
 
-long long Link_getBytesOutPerSecond(T L) {
+int64_t Link_getBytesOutPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->obytes)) : -1LL;
 }
 
 
-long long Link_getBytesOutPerMinute(T L, int count) {
+int64_t Link_getBytesOutPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->obytes), count) : -1LL;
 }
 
 
-long long Link_getBytesOutPerHour(T L, int count) {
+int64_t Link_getBytesOutPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->obytes), count) : -1LL;
 }
 
 
-long long Link_getBytesOutTotal(T L) {
+int64_t Link_getBytesOutTotal(T L) {
         assert(L);
         return L->state > 0 ? L->obytes.now : -1LL;
 }
@@ -467,49 +467,49 @@ double Link_getSaturationOutPerSecond(T L) {
 }
 
 
-long long Link_getPacketsOutPerSecond(T L) {
+int64_t Link_getPacketsOutPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->opackets)) : -1LL;
 }
 
 
-long long Link_getPacketsOutPerMinute(T L, int count) {
+int64_t Link_getPacketsOutPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->opackets), count) : -1LL;
 }
 
 
-long long Link_getPacketsOutPerHour(T L, int count) {
+int64_t Link_getPacketsOutPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->opackets), count) : -1LL;
 }
 
 
-long long Link_getPacketsOutTotal(T L) {
+int64_t Link_getPacketsOutTotal(T L) {
         assert(L);
         return L->state > 0 ? L->opackets.now : -1LL;
 }
 
 
-long long Link_getErrorsOutPerSecond(T L) {
+int64_t Link_getErrorsOutPerSecond(T L) {
         assert(L);
         return L->state > 0 ? _deltaSecond(L, &(L->oerrors)) : -1LL;
 }
 
 
-long long Link_getErrorsOutPerMinute(T L, int count) {
+int64_t Link_getErrorsOutPerMinute(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaMinute(L, &(L->oerrors), count) : -1LL;
 }
 
 
-long long Link_getErrorsOutPerHour(T L, int count) {
+int64_t Link_getErrorsOutPerHour(T L, int count) {
         assert(L);
         return L->state > 0 ? _deltaHour(L, &(L->oerrors), count) : -1LL;
 }
 
 
-long long Link_getErrorsOutTotal(T L) {
+int64_t Link_getErrorsOutTotal(T L) {
         assert(L);
         return L->state > 0 ? L->oerrors.now : -1LL;
 }
@@ -521,7 +521,7 @@ int Link_getState(T L) {
 }
 
 
-long long Link_getSpeed(T L) {
+int64_t Link_getSpeed(T L) {
         assert(L);
         return L->state > 0 ? L->speed : -1LL;
 }
